requirements.txt
anndata==0.11.4
array_api_compat==1.11.2
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.13.3
bioinfokit==2.1.4
biopython==1.85
blinker==1.9.0
blosc2==3.2.1
celltypist==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
clarabel==0.10.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cvxopt==1.3.2
cvxpy==1.6.2
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_cytoscape==1.0.2
dask==2025.2.0
debugpy==1.8.6
decorator==5.1.1
dnspython==2.7.0
docopt==0.6.2
dython==0.7.9
et_xmlfile==2.0.0
executing==2.1.0
fancyimpute==0.7.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.55.0
fsspec==2025.2.0
ftpretty==0.4.0
future==1.0.0
gdown==5.2.0
goatools==1.4.12
graphviz==0.20.3
GridDataFormats==1.0.2
gseapy==1.1.8
h5py==3.13.0
hyperopt==0.1.2
hypothesis==6.124.7
idna==3.10
igraph==0.11.8
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipymol==0.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kmapper==2.1.0
knnimpute==0.1.0
legacy-api-wrap==1.4.1
leidenalg==0.10.2
llvmlite==0.44.0
locket==1.0.0
logomaker==0.8.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
mda-xdrlib==0.2.0
mdahole2==0.5.0
MDAnalysis==2.8.0
MDAnalysisTests==2.8.0
mdurl==0.1.2
memory-profiler==0.61.0
mmtf-python==1.1.3
mpl-tools==0.4.0
mpmath==1.3.0
mrcfile==1.5.4
msgpack==1.1.0
narwhals==1.28.0
natsort==8.4.0
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==3.4.2
nose==1.3.7
numba==0.61.0
numexpr==2.10.2
numpy==2.1.2
openpyxl==3.1.5
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
pandas-flavor==0.6.0
parso==0.8.4
partd==1.4.2
pathsimanalysis==1.2.0
patsy==1.0.1
pillow==11.0.0
pingouin==0.5.5
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydot==3.0.4
Pygments==2.18.0
pymongo==4.11.3
pynndescent==0.5.13
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post5
qutip==5.1.1
qutip-qip==0.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rpds-py==0.24.0
scanpy==1.11.1
scikit-learn==1.5.2
scikit-misc==0.5.1
scipy==1.14.1
scs==3.2.7.post2
scvi==0.6.8
seaborn==0.13.2
sequence==0.7.0
session-info2==0.1.2
setuptools==75.8.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tables==3.10.2
tabulate==0.9.0
texttable==1.7.0
textwrap3==0.9.2
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ucsc-genomic-api==1.0.2
umap-learn==0.5.7
urllib3==2.3.0
waterdynamics==1.2.0
wcwidth==0.2.13
Werkzeug==3.0.6
xarray==2024.11.0
xgboost==3.0.0
xlrd==2.0.1
XlsxWriter==3.2.3
zipp==3.21.0
